+++ src/podofo/podofo.h
@@ -1,0 +1,19 @@
+/**
+ * SPDX-FileCopyrightText: (C) 2006 Dominik Seichter <domseichter@web.de>
+ * SPDX-FileCopyrightText: (C) 2021 Francesco Pretto <ceztko@gmail.com>
+ * SPDX-License-Identifier: LGPL-2.0-or-later
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+#ifndef PODOFO_H
+#define PODOFO_H
+
+/**
+ * This file can be used in client applications to include
+ * all files required by podofo at once.
+ */
+
+#include "podofo-base.h"
+#include "podofo-contrib.h"
+
+#endif // PODOFO_H