+++ test/unit/ColorTest.cpp
@@ -11,7 +11,7 @@
 #include <PdfTest.h>
 
 using namespace std;
-using namespace mm;
+using namespace PoDoFo;
 
 class TestColor
 {