+++ test/unit/ImageTest.cpp
@@ -9,7 +9,7 @@
 #include <PdfTest.h>
 
 using namespace std;
-using namespace mm;
+using namespace PoDoFo;
 
 TEST_CASE("TestImage1")
 {
@@ -55,12 +55,12 @@ TEST_CASE("TestImage2")
     TestUtils::SaveFramePPM(ppmbuffer, buffer.data(),
         PdfPixelFormat::BGRA, image->GetWidth(), image->GetHeight());
 
-#ifdef PDFMM_PLAYGROUND
+#ifdef PODOFO_PLAYGROUND
     // NOTE: The following check may file using different,
     // jpeg libraries such as libjpeg-turbo
     string expectedImage;
     TestUtils::ReadTestInputFileTo(expectedImage, "ReferenceImage.ppm");
 
     REQUIRE(ppmbuffer == expectedImage);
-#endif // PDFMM_PLAYGROUND
+#endif // PODOFO_PLAYGROUND
 }